+++ packages/live-preview-sdk/src/index.ts
@@ -1,5 +1,6 @@
-import { type DocumentNode } from 'graphql';
 import { encodeGraphQLResponse } from '@contentful/content-source-maps';
+import { type DocumentNode } from 'graphql';
+
 import { version } from '../package.json';
 import {
   debug,
@@ -307,5 +308,6 @@ export class ContentfulLivePreview {
 }
 
 export { LIVE_PREVIEW_EDITOR_SOURCE, LIVE_PREVIEW_SDK_SOURCE } from './constants';
+
 export * from './messages';
 export { encodeGraphQLResponse };