+++ packages/live-preview-sdk/src/liveUpdates.ts
@@ -34,34 +34,22 @@ export class LiveUpdates {
 
   /** Receives the data from the message event handler and calls the subscriptions */
   public async receiveMessage(message: MessageFromEditor): Promise<void> {
-    if (
-      ('action' in message && message.action === 'ENTRY_UPDATED') ||
-      message.method === LivePreviewPostMessageMethods.ENTRY_UPDATED
-    ) {
-      const { data } = message as EntryUpdatedMessage;
+    if (message.method === LivePreviewPostMessageMethods.ENTRY_UPDATED) {
+      const { data, subscriptionId } = message as EntryUpdatedMessage;
 
-      await Promise.all(
-        [...this.subscriptions].map(async ([, s]) => {
-          try {
-            // Only if there was an update, trigger the callback to avoid unnecessary re-renders
-            if (s.sysId === data.sys.id) {
-              s.callback(data);
-            }
-          } catch (error) {
-            this.sendErrorMessage({
-              message: (error as Error).message,
-              payload: { data: s.data, update: data },
-              type: 'SUBSCRIPTION_UPDATE_FAILED',
-            });
+      const subscription = this.subscriptions.get(subscriptionId);
 
-            debug.error('Failed to apply live update', {
-              error,
-              subscribedData: s.data,
-              updateFromEditor: data,
-            });
-          }
-        })
-      );
+      if (subscription) {
+        subscription.callback(data);
+        subscription.data = data;
+        this.subscriptions.set(subscriptionId, subscription);
+      } else {
+        debug.error('Received an update for an unknown subscription', {
+          subscriptionId,
+          data,
+          subscriptions: this.subscriptions,
+        });
+      }
     }
   }
 
@@ -110,13 +98,13 @@ export class LiveUpdates {
    * Will be called once initially for the restored data
    */
   public subscribe(originalConfig: ContentfulSubscribeConfig): VoidFunction {
-    const { isGQL, isValid, sysId, isREST, config } =
+    const { isGQL, isValid, sysIds, isREST, config } =
       validateLiveUpdatesConfiguration(originalConfig);
 
     if (!isValid) {
       this.sendErrorMessage({
         message: 'Failed to subscribe',
-        payload: { isGQL, isValid, sysId, isREST },
+        payload: { isGQL, isValid, sysIds, isREST },
         type: 'SUBSCRIPTION_SETUP_FAILED',
       });
       return () => {
@@ -129,7 +117,7 @@ export class LiveUpdates {
 
     this.subscriptions.set(id, {
       ...config,
-      sysId,
+      sysIds,
       gqlParams: config.query ? parseGraphQLParams(config.query) : undefined,
     });
 
@@ -146,7 +134,7 @@ export class LiveUpdates {
     const message: Omit<SubscribedMessage, 'action'> | UnsubscribedMessage = {
       type: isGQL ? 'GQL' : 'REST',
       locale,
-      entryId: sysId,
+      sysIds,
       event: 'edit',
       id,
       config: stringify(config),